engine/schema: add empty DB upgrade path from 4.15 to 4.16
