+++ engine/schema/src/main/java/com/cloud/upgrade/DatabaseUpgradeChecker.java
@@ -69,6 +69,8 @@ import com.cloud.upgrade.dao.Upgrade41200to41300;
 import com.cloud.upgrade.dao.Upgrade41300to41310;
 import com.cloud.upgrade.dao.Upgrade41310to41400;
 import com.cloud.upgrade.dao.Upgrade41400to41500;
+import com.cloud.upgrade.dao.Upgrade41500to41510;
+import com.cloud.upgrade.dao.Upgrade41510to41600;
 import com.cloud.upgrade.dao.Upgrade420to421;
 import com.cloud.upgrade.dao.Upgrade421to430;
 import com.cloud.upgrade.dao.Upgrade430to440;
@@ -193,6 +195,8 @@ public class DatabaseUpgradeChecker implements SystemIntegrityChecker {
                 .next("4.13.0.0", new Upgrade41300to41310())
                 .next("4.13.1.0", new Upgrade41310to41400())
                 .next("4.14.0.0", new Upgrade41400to41500())
+                .next("4.15.0.0", new Upgrade41500to41510())
+                .next("4.15.1.0", new Upgrade41510to41600())
                 .build();
     }
